wip: pointwise limits
The end goal is to prove that `Tm` in CCC.Free is cartesian by showing
that it's the composition of two cartesian functors (this makes
the rest of the module prohibitively slow, so we probably want to keep
the direct definition).

Thus I need to show that よ and precompose F preserve limits, which
requires talking about pointwise limits.

In order to show that limits in a functor category with complete
codomain are computed pointwise, I thought I would show that:
- Limits of the form 1 ← C → [D, E] are equivalently Kan extensions of
  the form D ← C × D → E by currying, and
- those limits are pointwise exactly when the corresponding Kan
  extension is pointwise.

However, this requires generalising the universe levels in
Kan.Pointwise, which is very annoying.